feature/4091: Refactors production and revenue tabs
