+++ frontend-nuxt/app/components/revenue-data/RevenueTab.vue
@@ -1,0 +1,7 @@
+<script setup>
+</script>
+
+<template>
+  <h3>Revenue</h3>
+  <p>The amount of money collected from natural resources on property owned by the federal government and Native Americans.</p>
+</template>
